refactor: Extract sound_effect_util.py as a single service module.
